requirements.txt
celery
dnspython
django==2.2.3
django-allauth
django-countries
django-subscriptions
django-celery-beat
django-celery-results
django-celery-email
django-widget-tweaks
django-recaptcha
dj-stripe
gunicorn
mysqlclient
pyopenssl
quickemailverification
redis
tldextract
unicodecsv
validators
